data-product-download/requirements.txt
tqdm==4.65.0
python-dotenv==1.0.0
requests==2.28.1
click==8.1.3